tests/wpt/tests/css/css-overflow/parsing/scroll-markers-invalid.tentative.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title>CSS Overflow: parsing scroll-markers with invalid values</title>
  <link rel="help" href="https://github.com/w3c/csswg-drafts/pull/10243">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>

<body>
  <div id="target"></div>
  <script>
    test_invalid_value('scroll-markers', '10');
    test_invalid_value('scroll-markers', 'true');
    test_invalid_value('scroll-markers', 'default');
    test_invalid_value('scroll-markers', 'set');
    test_invalid_value('scroll-markers', 'before, after');
  </script>
</body>

</html>
